requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
babel==2.13.1
    # via tkcalendar
bottle==0.12.25
    # via pywebview
certifi==2022.12.7
    # via requests
cffi==1.16.0
    # via clr-loader
chardet==3.0.4
    # via requests
click==8.1.7
    # via nltk
clr-loader==0.2.6
    # via pythonnet
colorama==0.4.6
    # via
    #   click
    #   tqdm
customtkinter==5.2.1
    # via -r requirements.in
darkdetect==0.8.0
    # via customtkinter
docx2pdf==0.1.8
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
idna==2.10
    # via requests
joblib==1.3.2
    # via nltk
language-tool-python==2.7.1
    # via -r requirements.in
lxml==4.9.3
    # via python-docx
nltk==3.8.1
    # via -r requirements.in
numpy==1.26.2
    # via pandas
openpyxl==3.1.2
    # via -r requirements.in
pandas==2.1.3
    # via -r requirements.in
proxy-tools==0.1.0
    # via pywebview
pycparser==2.21
    # via cffi
pymysql==1.1.0
    # via -r requirements.in
pypdf==3.17.3
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
python-docx==1.1.0
    # via -r requirements.in
pythonnet==3.0.3
    # via pywebview
pytz==2023.3.post1
    # via pandas
pywebview==4.4.1
    # via -r requirements.in
pywin32==306
    # via docx2pdf
regex==2023.10.3
    # via nltk
requests==2.24.0
    # via language-tool-python
six==1.16.0
    # via python-dateutil
termcolor==2.4.0
    # via -r requirements.in
tk==0.1.0
    # via -r requirements.in
tkcalendar==1.6.1
    # via -r requirements.in
tkinter-tooltip==2.1.0
    # via -r requirements.in
tqdm==4.66.1
    # via
    #   docx2pdf
    #   language-tool-python
    #   nltk
typing-extensions==4.8.0
    # via
    #   python-docx
    #   pywebview
tzdata==2023.3
    # via pandas
urllib3==1.25.11
    # via requests
